chore: move test for context
